chore: move legacy demo scenes to examples + tests for integrator/tangent/shape variants
