+++ crates/rye-math/src/tangent.rs
@@ -61,3 +61,50 @@ where
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::EuclideanR3;
+    use glam::Vec3;
+
+    #[test]
+    fn new_stores_at_and_v_directly() {
+        let t = Tangent::<EuclideanR3>::new(Vec3::X, Vec3::Y);
+        assert_eq!(t.at, Vec3::X);
+        assert_eq!(t.v, Vec3::Y);
+    }
+
+    /// `exp` walks from `at` along `v` for unit time. In E³ that's
+    /// just `at + v`, the trivial path that pins the contract.
+    #[test]
+    fn exp_in_flat_space_is_at_plus_v() {
+        let t = Tangent::<EuclideanR3>::new(Vec3::new(1.0, 2.0, 3.0), Vec3::new(0.5, 0.0, -0.5));
+        assert_eq!(t.exp(&EuclideanR3), Vec3::new(1.5, 2.0, 2.5));
+    }
+
+    /// `scale` rescales `v` and leaves `at` alone; round-tripping by
+    /// the inverse scale must restore the original vector exactly in
+    /// flat space.
+    #[test]
+    fn scale_preserves_base_and_is_invertible() {
+        let original = Tangent::<EuclideanR3>::new(Vec3::Y, Vec3::new(2.0, 3.0, 4.0));
+        let scaled = original.scale(2.5);
+        assert_eq!(scaled.at, Vec3::Y);
+        assert_eq!(scaled.v, Vec3::new(5.0, 7.5, 10.0));
+        let back = scaled.scale(1.0 / 2.5);
+        assert!((back.v - original.v).length() < 1e-6);
+    }
+
+    /// `transport_to` in E³ keeps `v` unchanged (parallel transport
+    /// is the identity in flat space) but updates `at`. Pins that
+    /// the path-aware primitive doesn't accidentally rotate the
+    /// vector in the trivial case.
+    #[test]
+    fn transport_to_in_flat_space_preserves_vector() {
+        let t = Tangent::<EuclideanR3>::new(Vec3::ZERO, Vec3::new(1.0, 2.0, 3.0));
+        let moved = t.transport_to(&EuclideanR3, Vec3::new(5.0, 0.0, 0.0));
+        assert_eq!(moved.at, Vec3::new(5.0, 0.0, 0.0));
+        assert_eq!(moved.v, Vec3::new(1.0, 2.0, 3.0));
+    }
+}
